+++ src/Order/Changelog.md
@@ -2,8 +2,8 @@
 
 ## 3.x Series
 
-## Unreleased
-##### 2023-XX-YY
+## 3.6.1
+##### 2023-03-09
 
 - Fixed incorrect zone lists due to a bug in Address module v2.5.0 (bump to v2.5.1)